+++ SoundFontInfoLib/SoundFontInfoLib/Entity/SampleHeader.hpp
@@ -76,10 +76,10 @@ public:
   size_t sampleRate() const { return dwSampleRate; }
   
   /// @returns the MIDI key (frequency) for the source samples
-  short originalMIDIKey() const { return originalKey; }
+  int originalMIDIKey() const { return originalKey; }
   
   /// @returns the pitch correction to apply when playing back the samples
-  short pitchCorrection() const { return correction; }
+  int pitchCorrection() const { return correction; }
   
   void dump(const std::string& indent, size_t index) const;
